Introduce Strongly-typed wrapper for Offers Ledger Object
This commit defines and uses a new wrapper for the Offers Ledger Object. It is a sequel to the continuing work in the series of Account, DepositPreauth and Checks.
Existing usages of readSLE and peekSLE relevant to the Offers object have been modified to use read and peek functions respectively. This brings forth the compile-time type checking advantages offered by the proposed Strongly typed wrapper.

As part of future work, DeleteAccount.cpp needs to be updated to ensure that std::shared_ptr<SLE> is removed in favor of the new wrapper. This requires changes to other ledger objects like NFTokens.