Simplified function signatures for read, peek functions
This commit proposes a single template parameter for the ReadView::read() and ApplyView::peek() functions.
The former returns a non-writable SLE whereas the latter returns a writable version of SLE. I have not found any instances to the contrary.

Furthermore, the auto return type can be replaced with the right-hand-side version of -> in the function signature. This commit builds properly and passes existing unit tests.